Move bulk operation into indexer namespace
