+++ lib/Alchemy/Phrasea/SearchEngine/Elastic/Indexer/BulkOperation.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Alchemy\Phrasea\SearchEngine\Elastic;
+namespace Alchemy\Phrasea\SearchEngine\Elastic\Indexer;
 
 use Alchemy\Phrasea\SearchEngine\Elastic\Exception\Exception;
 use Elasticsearch\Client;
